backend/requirements.txt
Flask
paho-mqtt
pymongo
python-dotenv